+++ host/class/uvc/usb_host_uvc/uvc_control.c
@@ -86,27 +86,27 @@ static esp_err_t uvc_host_stream_control(uvc_host_stream_hdl_t stream_hdl, uvc_v
     return ret;
 }
 
-static inline esp_err_t uvc_host_stream_control_probe_set(uvc_host_stream_hdl_t stream_hdl, uvc_vs_ctrl_t *vs_control, const uvc_host_stream_format_t *vs_format)
+static inline esp_err_t uvc_control_probe_set(uvc_host_stream_hdl_t stream_hdl, uvc_vs_ctrl_t *vs_control, const uvc_host_stream_format_t *vs_format)
 {
     return uvc_host_stream_control(stream_hdl, vs_control, (uvc_host_stream_format_t *)vs_format, UVC_SET_CUR, false);
 }
 
-static inline esp_err_t uvc_host_stream_control_probe_get(uvc_host_stream_hdl_t stream_hdl, uvc_vs_ctrl_t *vs_control, uvc_host_stream_format_t *vs_format)
+static inline esp_err_t uvc_control_probe_get(uvc_host_stream_hdl_t stream_hdl, uvc_vs_ctrl_t *vs_control, uvc_host_stream_format_t *vs_format)
 {
     return uvc_host_stream_control(stream_hdl, vs_control, vs_format, UVC_GET_CUR, false);
 }
 
-static inline esp_err_t uvc_host_stream_control_probe_get_max(uvc_host_stream_hdl_t stream_hdl, uvc_vs_ctrl_t *vs_control, uvc_host_stream_format_t *vs_format)
+static inline esp_err_t uvc_control_probe_get_max(uvc_host_stream_hdl_t stream_hdl, uvc_vs_ctrl_t *vs_control, uvc_host_stream_format_t *vs_format)
 {
     return uvc_host_stream_control(stream_hdl, vs_control, vs_format, UVC_GET_MAX, false);
 }
 
-static inline esp_err_t uvc_host_stream_control_probe_get_min(uvc_host_stream_hdl_t stream_hdl, uvc_vs_ctrl_t *vs_control, uvc_host_stream_format_t *vs_format)
+static inline esp_err_t uvc_control_probe_get_min(uvc_host_stream_hdl_t stream_hdl, uvc_vs_ctrl_t *vs_control, uvc_host_stream_format_t *vs_format)
 {
     return uvc_host_stream_control(stream_hdl, vs_control, vs_format, UVC_GET_MIN, false);
 }
 
-static inline esp_err_t uvc_host_stream_control_commit(uvc_host_stream_hdl_t stream_hdl, uvc_vs_ctrl_t *vs_control, const uvc_host_stream_format_t *vs_format)
+static inline esp_err_t uvc_control_commit(uvc_host_stream_hdl_t stream_hdl, uvc_vs_ctrl_t *vs_control, const uvc_host_stream_format_t *vs_format)
 {
     return uvc_host_stream_control(stream_hdl, vs_control, (uvc_host_stream_format_t *)vs_format, UVC_SET_CUR, true);
 }
@@ -122,7 +122,28 @@ static inline bool uvc_is_vs_format_equal(const uvc_host_stream_format_t *a, con
     return false;
 }
 
-esp_err_t uvc_host_stream_control_negotiate(uvc_host_stream_hdl_t stream_hdl, const uvc_host_stream_format_t *vs_format, uvc_vs_ctrl_t *vs_result_ret)
+esp_err_t uvc_host_stream_control_probe(uvc_host_stream_hdl_t stream_hdl, const uvc_host_stream_format_t *vs_format, uvc_vs_ctrl_t *vs_result_ret)
+{
+    UVC_CHECK(stream_hdl && vs_format, ESP_ERR_INVALID_ARG);
+
+    esp_err_t ret;
+    uvc_vs_ctrl_t vs_result = {0};
+    uvc_host_stream_format_t format_ignored;
+
+    // Notes: Some cameras require 'probe_get' to be the 1st negotiation call
+    uvc_control_probe_get(stream_hdl, &vs_result, &format_ignored);
+    uvc_control_probe_set(stream_hdl, &vs_result, vs_format);       // Set the desired frame format
+    ret = uvc_control_probe_get(stream_hdl, &vs_result, &format_ignored); // Get back the format (not checked yet)
+
+    // Pass the result to user
+    if (vs_result_ret) {
+        memcpy(vs_result_ret, &vs_result, sizeof(uvc_vs_ctrl_t));
+    }
+
+    return ret;
+}
+
+esp_err_t uvc_host_stream_control_commit(uvc_host_stream_hdl_t stream_hdl, const uvc_host_stream_format_t *vs_format)
 {
     /*
     This is not a 'real' negotiation as we return OK only if set the format exactly as expected.
@@ -139,18 +160,16 @@ esp_err_t uvc_host_stream_control_negotiate(uvc_host_stream_hdl_t stream_hdl, co
         // Notes: Some cameras require 'probe_get' to be the 1st negotiation call
         // It returns 'default' format and frame settings. It is ignored for now.
         // We can reuse the returned value, if we wanted to implement 'negotiate default frame format' feature.
-        uvc_host_stream_control_probe_get(stream_hdl, &vs_result, &format_ignored);
-        uvc_host_stream_control_probe_set(stream_hdl, &vs_result, vs_format);       // Set the desired frame format
-        uvc_host_stream_control_probe_get(stream_hdl, &vs_result, &format_ignored); // Get back the format (not checked yet)
+        uvc_host_stream_control_probe(stream_hdl, vs_format, NULL);
 
         // We do this to mimic Windows driver: The Min/Max values are ignored by this driver.
         // These values can be used if we wanted to negotiate advanced parameters, such as wCompQuality to select JPEG encoding quality
-        uvc_host_stream_control_probe_get_max(stream_hdl, &vs_result_ignored, &format_ignored);
-        uvc_host_stream_control_probe_get_min(stream_hdl, &vs_result_ignored, &format_ignored);
+        uvc_control_probe_get_max(stream_hdl, &vs_result_ignored, &format_ignored);
+        uvc_control_probe_get_min(stream_hdl, &vs_result_ignored, &format_ignored);
 
         // Probe that the camera accepts our format before committing
-        ret = uvc_host_stream_control_probe_set(stream_hdl, &vs_result, vs_format);
-        ret |= uvc_host_stream_control_probe_get(stream_hdl, &vs_result, &format_set);
+        ret = uvc_control_probe_set(stream_hdl, &vs_result, vs_format);
+        ret |= uvc_control_probe_get(stream_hdl, &vs_result, &format_set);
         UVC_CHECK(ret == ESP_OK, ret);
         if (uvc_is_vs_format_equal(&format_set, vs_format)) {
             // If the 'set format' equals 'get format', the camera accepts our format and we can commit it
@@ -162,12 +181,5 @@ esp_err_t uvc_host_stream_control_negotiate(uvc_host_stream_hdl_t stream_hdl, co
              vs_format->h_res, vs_format->v_res, vs_format->fps, format_set.h_res, format_set.v_res, format_set.fps);
 
     // Commit the negotiated format
-    ret = uvc_host_stream_control_commit(stream_hdl, &vs_result, vs_format);
-
-    // Pass the result to user
-    if (vs_result_ret) {
-        memcpy(vs_result_ret, &vs_result, sizeof(uvc_vs_ctrl_t));
-    }
-
-    return ret;
+    return uvc_control_commit(stream_hdl, &vs_result, vs_format);
 }