+++ host/class/uvc/usb_host_uvc/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - Added `uvc_host_stream_format_select()` function that allows change of format of an opened stream
+- Fixed device opening procedure. Now the frame format is committed when the stream is started
 
 ## 2.1.0